+++ lib/common/globals.h
@@ -46,7 +46,6 @@ DEFINE_LIST_WITH_DTOR(show_boxes, char*, free)
     GLOBALS_API EXTERN char *Version;
     GLOBALS_API EXTERN char **Files;	/* from command line */
     GLOBALS_API EXTERN const char **Lib;		/* from command line */
-    GLOBALS_API EXTERN char *CmdName;
     GLOBALS_API EXTERN char *Gvfilepath;  /* Per-process path of files allowed in image attributes (also ps libs) */
     GLOBALS_API EXTERN char *Gvimagepath; /* Per-graph path of files allowed in image attributes  (also ps libs) */